+++ test/authorizer/none-authorizer.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { getMockedRequest, mockConfig, mockFetch } from './authorizer-mocks.js';
+import { getMockedRequest, mockConfig, mockFetchUser } from './authorizer-mocks.js';
 
 vi.mock('../../src/config.js', () => {
   return mockConfig({
@@ -9,7 +9,7 @@ vi.mock('../../src/config.js', () => {
 
 vi.mock('node-fetch', () => {
   return {
-    default: mockFetch
+    default: mockFetchUser
   };
 });
 
@@ -17,12 +17,6 @@ import { getAuthorizer } from '../../src/authorizer.js';
 const authorizer = getAuthorizer();
 
 describe('None authorizer', () => {
-  it('Anonymous user with valid relative path', async () => {
-    const request = getMockedRequest('/foo/bar', undefined);
-    const path = await authorizer.getAuthorizedPath(request);
-    expect(path).eq('/path/to/zarr/data/foo/bar');
-  });
-
   it('Registered user with valid absolute path', async () => {
     const request = getMockedRequest('/path/to/zarr/data/foo/bar', 'cookie-user-1');
     const path = await authorizer.getAuthorizedPath(request);
@@ -33,5 +27,11 @@ describe('None authorizer', () => {
     const request = getMockedRequest('../foo/bar', 'cookie-user-1');
     const path = await authorizer.getAuthorizedPath(request);
     expect(path).eq(undefined);
+  });
+
+  it('Path with URL encoded characters', async () => {
+    const request = getMockedRequest('/path/to/zarr/data/foo/bar%23baz', undefined);
+    const path = await authorizer.getAuthorizedPath(request);
+    expect(path).eq('/path/to/zarr/data/foo/bar#baz');
   });
 });