+++ test/authorizer/user-folders-authorizer.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { getMockedRequest, mockConfig, mockFetch } from './authorizer-mocks.js';
+import { getMockedRequest, mockConfig, mockFetchUser } from './authorizer-mocks.js';
 
 vi.mock('../../src/config.js', () => {
   return mockConfig({
@@ -9,8 +9,36 @@ vi.mock('../../src/config.js', () => {
 
 vi.mock('node-fetch', () => {
   return {
-    default: mockFetch
-  };
+    default: function (path: string, { headers }) {
+      if (path.endsWith('/auth/current-user/settings/')) {
+        const cookie = headers['Cookie'];
+        switch (cookie) {
+          case 'cookie-user-1':
+            return {
+              ok: true,
+              json: () => ({ slurm_user: 'admin' })
+            };
+          case 'cookie-user-2':
+            return {
+              ok: true,
+              json: () => ({ slurm_user: 'user2' })
+            };
+          case 'cookie-user-3':
+            return {
+              ok: false,
+              json: () => ({ detail: 'error' })
+            };
+          default:
+            return {
+              ok: false,
+              json: () => ({})
+            };
+        }
+      } else {
+        return mockFetchUser(path, { headers });
+      }
+    }
+  }
 });
 
 import { getAuthorizer } from '../../src/authorizer.js';
@@ -23,26 +51,20 @@ describe('User folders authorizer', () => {
     expect(path).eq('/path/to/zarr/data/admin/foo/bar');
   });
 
-  it('Registered user2 with valid relative path', async () => {
-    const request = getMockedRequest('/user2/foo/bar', 'cookie-user-2');
-    const path = await authorizer.getAuthorizedPath(request);
-    expect(path).eq('/path/to/zarr/data/user2/foo/bar');
-  });
-
   it('Registered user with path of another user', async () => {
     const request = getMockedRequest('/path/to/zarr/data/admin/foo/bar', 'cookie-user-2');
     const path = await authorizer.getAuthorizedPath(request);
     expect(path).eq(undefined);
   });
 
-  it('Anonymous user with valid relative path', async () => {
-    const request = getMockedRequest('/foo/bar', undefined);
+  it('Registered user with invalid path', async () => {
+    const request = getMockedRequest('../foo/bar', 'cookie-user-1');
     const path = await authorizer.getAuthorizedPath(request);
     expect(path).eq(undefined);
   });
 
-  it('Registered user with invalid path', async () => {
-    const request = getMockedRequest('../foo/bar', 'cookie-user-1');
+  it('/auth/current-user/settings/ returns error', async () => {
+    const request = getMockedRequest('/user2/foo/bar', 'cookie-user-3');
     const path = await authorizer.getAuthorizedPath(request);
     expect(path).eq(undefined);
   });