+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 Note: Numbers like (#123) point to closed Pull Requests on the fractal-vizarr-viewer repository.
 
+# Unreleased
+
+* Supported `AUTHORIZATION_SCHEME="fractal-server-viewer-paths"` (\#35);
+* Supported URL encoded characters in path (\#35);
+* Deprecated `AUTHORIZATION_SCHEME="allowed-list"` (\#35);
+* Removed support for relative paths (\#35);
+
 # 0.1.4
 
 * Fix broken `AUTHORIZATION_SCHEME=user-folders` (\#34);